refactor: Reorganize workflow resource imports
