+++ packages/framework/src/index.ts
@@ -1,5 +1,5 @@
 export { Client } from './client';
-export * from './types';
 export * from './constants';
 export { NovuRequestHandler, type ServeHandlerOptions } from './handler';
-export { workflow } from './workflow';
+export * from './resources';
+export * from './types';
