+++ packages/framework/src/resources/workflow.test.ts
@@ -1,5 +1,5 @@
 import { it, describe, beforeEach, expect, vi, afterEach } from 'vitest';
-import { MissingSecretKeyError } from './errors';
+import { MissingSecretKeyError } from '../errors';
 import { workflow } from './workflow';
 
 describe('workflow function', () => {
